+++ NEWS.md
@@ -12,6 +12,8 @@
 * New `makeOxfordPlots()` function to create macrosynteny plots.  Thanks to
   Aleksandra Bliznina.
 * A new set of example 5-uplets, to illustrate double inversions.
+* Flag double inversions with `flagDoubleInversions()` and `flagAll()`, test
+  them with `exampleDoubleInversion`.
 
 ## Bug fixes